update source URL; remove uncessary format_string() call; v0.2.3
I noticed that I had the wrong URL set for the Source and Issues to be displayed
in the Chef Supermarket UI. Oops. :(

This also makes a change to the `command_string` helper method, in that it
removes an unneeded invocation of `format_string()` when using the resource name
as the label.

Finally, bump to v0.2.3 for release.